+++ sdk/monitor/monitor-opentelemetry-exporter/src/platform/nodejs/index.ts
@@ -6,5 +6,6 @@
  */
 export * from "./constants.js";
 export * from "./persist/index.js";
+export * from "./baseSender.js";
 export * from "./httpSender.js";
 export * from "./context/index.js";